Sensors/Imagenex881A: Fixed conversion error.

+++ src/Sensors/Imagenex881A/Parser.hpp
@@ -230,7 +230,7 @@ namespace Sensors
       {
         uint8_t hb = (hbyte & 0x7e) >> 1;
         uint8_t lb = (((hbyte & 0x01) << 7) | (lbyte & 0x7f));
-        uint16_t r = (hb << 8) | lb;
+        float r = (float)((hb << 8) | lb);
 
         if (m_range < 5)
           m_profile_range = (r * 2) / 1000;
